Removed the unneeded interfaces
