+++ src/main/java/com/sethhaskellcondie/thegamepensiveapi/domain/system/SystemGateway.java
@@ -1,6 +1,14 @@
 package com.sethhaskellcondie.thegamepensiveapi.domain.system;
 
 import com.sethhaskellcondie.thegamepensiveapi.domain.EntityGateway;
+import org.springframework.stereotype.Component;
 
-public interface SystemGateway extends EntityGateway<System, SystemRequestDto, SystemResponseDto> {
+import com.sethhaskellcondie.thegamepensiveapi.domain.EntityGatewayAbstract;
+import com.sethhaskellcondie.thegamepensiveapi.domain.EntityService;
+
+@Component
+public class SystemGateway extends EntityGatewayAbstract<System, SystemRequestDto, SystemResponseDto> implements EntityGateway<System, SystemRequestDto, SystemResponseDto> {
+    public SystemGateway(EntityService<System, SystemRequestDto, SystemResponseDto> service) {
+        super(service);
+    }
 }
