--- src/main/java/com/sethhaskellcondie/thegamepensiveapi/domain/system/SystemGatewayImpl.java
@@ -1,13 +1,0 @@
-package com.sethhaskellcondie.thegamepensiveapi.domain.system;
-
-import org.springframework.stereotype.Component;
-
-import com.sethhaskellcondie.thegamepensiveapi.domain.EntityGatewayImpl;
-import com.sethhaskellcondie.thegamepensiveapi.domain.EntityService;
-
-@Component
-public class SystemGatewayImpl extends EntityGatewayImpl<System, SystemRequestDto, SystemResponseDto> implements SystemGateway {
-    public SystemGatewayImpl(EntityService<System, SystemRequestDto, SystemResponseDto> service) {
-        super(service);
-    }
-}
