--- src/main/java/com/sethhaskellcondie/thegamepensiveapi/domain/toy/ToyGatewayImpl.java
@@ -1,13 +1,0 @@
-package com.sethhaskellcondie.thegamepensiveapi.domain.toy;
-
-import org.springframework.stereotype.Component;
-
-import com.sethhaskellcondie.thegamepensiveapi.domain.EntityGatewayImpl;
-import com.sethhaskellcondie.thegamepensiveapi.domain.EntityService;
-
-@Component
-public class ToyGatewayImpl extends EntityGatewayImpl<Toy, ToyRequestDto, ToyResponseDto> implements ToyGateway {
-    public ToyGatewayImpl(EntityService<Toy, ToyRequestDto, ToyResponseDto> service) {
-        super(service);
-    }
-}
